chore: move zod.data.ts to data dir and fixes type error
